+++ aws-ecs-github-deploy-policy/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Default `account_id` to the AWS provider account.
+
 ## 0.1.0 (2026-08-21)
 
 - Initial release.